i just area square and rectangle and need to fix triangle

+++ 1-Computers-and-Programs/imports.py
@@ -4,5 +4,5 @@ x = input("enter any number")
 y = math.sqrt(int(x))
 a = math.pow(int(x),2)
 
-print("Square Root value=",float(y))
-print("square value=",float(a))
+print("Square Root value=",int(y))
+print("square value=",int(a))
